# GGRC-6911 Add benchmarks for export functionality